+++ peglib-core/src/test/java/org/pragmatica/peg/perf/CorpusParityTest.java
@@ -54,7 +54,7 @@ class CorpusParityTest {
     void cstHashMatchesBaseline(Path file) throws Exception {
         String source = Files.readString(file, StandardCharsets.UTF_8);
         var cst = GeneratedJava25Parser.parseToCst(source);
-        String actual = CstHash.of(cst);
+        String actual = CstHash.cstHash(cst);
 
         Path relative = CORPUS_ROOT.relativize(file);
         Path hashFile = BASELINE_ROOT.resolve(relative + ".hash");