+++ peglib-core/src/test/java/org/pragmatica/peg/perf/Phase1ParityTest.java
@@ -62,7 +62,7 @@ class Phase1ParityTest {
     void phase1OptimizedCstHashMatchesBaseline(Path file) throws Exception {
         String source = Files.readString(file, StandardCharsets.UTF_8);
         var cst = GeneratedJava25Parser.parseToCst(source, PHASE1_ALL_ON);
-        String actual = CstHash.of(cst);
+        String actual = CstHash.cstHash(cst);
 
         Path relative = CORPUS_ROOT.relativize(file);
         Path hashFile = BASELINE_ROOT.resolve(relative + ".hash");