+++ peglib-core/src/test/java/org/pragmatica/peg/perf/Phase2InlineLocationsParityTest.java
@@ -65,7 +65,7 @@ class Phase2InlineLocationsParityTest {
     void phase2InlineLocationsCstHashMatchesBaseline(Path file) throws Exception {
         String source = Files.readString(file, StandardCharsets.UTF_8);
         var cst = GeneratedJava25Parser.parseToCst(source, PHASE2_INLINE_LOCATIONS_ON);
-        String actual = CstHash.of(cst);
+        String actual = CstHash.cstHash(cst);
 
         Path relative = CORPUS_ROOT.relativize(file);
         Path hashFile = BASELINE_ROOT.resolve(relative + ".hash");